backend/requirements.txt
chromadb==1.5.5
fastapi==0.135.3
numpy==2.4.4
ollama==0.6.1
pydantic==2.12.5
pymupdf==1.27.2.2
python_docx==1.2.0
python-multipart==0.0.22
